chore: revert swapping help and description of presentation-role-conflict (#4738)
https://github.com/dequelabs/axe-core/pull/4685 made a change that
swapped the help and description text of `presentation-role-conflict`,
probably in order to put the "Ensure" text as the description (which is
what the pr updated). However the description should always been the
longer of the text and the help text short.

This reverts that change, but updates the description and help text to
bring it more in line with the "Ensure" pattern of the description.