chore: refactor away from looking at role attribute (#4756)
This work is in preparation to support fallback roles. Any part of the
code that used `.getAttribute('role')` or `.attr('role')` was migrated
to use `getEplicitRole()` instead. There are a few places that still
look at the role attribute directly, but those cases are validating the
attributes value for different things so it's a valid use case (e.g.
looking for abstract roles, invalid roles, fallback roles, unallowed
roles, etc.).

Also migrates some of the older test code to use arrow functions and
`const` rather than `var` (which is a tech debt strategy to slowly
update the code as we touch the files).

Part of https://github.com/dequelabs/axe-core/issues/3768

---------

Co-authored-by: Wilco Fiers <WilcoFiers@users.noreply.github.com>