fix(performanceTimer): throwing in axe catch clause (#4852)
I ran into a problem with this:


https://github.com/dequelabs/axe-core/blob/e7dae4ec48cbfef74de9f833fdcfb178c1002985/lib/core/base/rule.js#L297-L300

If the end mark isn't set, performance.measure throws an error. Since we
call performance.measure in the catch, it is quite possible that the end
mark doesn't exist. And then throwing from a catch completely throws the
run. I don't think axe should fail if the performance measure fails, so
instead I'm going to have it just log the error and move forward.