refactor: refactor flatten-tree to use return early style (#4854)
Was making changes to this file and found it hard to read through with
the multiple nesting of if statements and the reduces. Refactored to use
early return style to prevent the nesting and moved away from
array.reduce to just use a forEach over the array to add them (as well
as a function to do this since we do it multiple times).

---------

Co-authored-by: Wilco Fiers <WilcoFiers@users.noreply.github.com>