test(performance-timer): reduce flakiness of tests involving actual sleeps (#4890)
Fixes some flaky test cases that were failing as often as not in the
`test_firefox` job, where we were seeing in practice that `await
sleep(100)` was frequently taking much longer than 100ms to actually
complete. This updates the tests to avoid using "closeTo" in assertions
of how long a sleep lasts, in favor of `atLeast` + verifying nearness to
what `performance.now()` reports.

I verified this in 10x runs against firefox locally (where it exhibited
the same failure before the fix and no failure after the fix) + 3x runs
in CI:
- [test_firefox job
79723](https://app.circleci.com/pipelines/github/dequelabs/axe-core/7408/workflows/2240fff4-dd89-4b20-9a45-8865cd1a4710/jobs/79723)
- [test_firefox job
79734](https://app.circleci.com/pipelines/github/dequelabs/axe-core/7409/workflows/42c46140-8294-48c7-a8d2-b5261c25f585/jobs/79734)
- [test_firefox job
79743](https://app.circleci.com/pipelines/github/dequelabs/axe-core/7409/workflows/7cf713ff-d153-4de3-ade8-e0379553bd23/jobs/79743)